+++ CHANGELOG.md
@@ -44,6 +44,9 @@
   - fix SPIR-V generation from WGSL, which was broken due to "Kernel" capability
   - validate buffer storage classes
 
+## Unreleased
+  - Added support for storage texture arrays for Vulkan and Metal.
+
 ## v0.8 (2021-04-29)
   - Naga is used by default to translate shaders, SPIRV-Cross is optional behind `cross` feature
   - Features: